users authentiaction and authorization through api done

+++ controllers/api/v1/posts_api.js
@@ -23,17 +23,18 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
         if(post){
-            // if(req.user.id == post.user){
+            if(req.user.id == post.user){
                 post.remove();
                 await Comments.deleteMany({ post: req.params.id});
                
                 return res.json(200, {
                     message: 'posts and associated comments deleted successfully'
                 })
-        //     }
-        // }else{
-        //     req.flash('error', 'You cannot delete this post');
-        //     res.redirect('back');
+            }
+        }else{
+            return res.json(401, {
+                message: 'You cannot delete this post'
+            })
         }
     }catch(err){
         return res.status(500).json({
